refactor(l1): remove usize usage from constants in network crate (#4116)
**Motivation**

Avoid unnecessary usage of `usize` in the codebase.

**Description**

The approach taken was the following: if the constant is used only for
indexing or to be compared against `usize` values, keep it as `usize`.
Otherwise, change it to `u32` or `u64`, depending on the case.

Related to #4081

+++ crates/networking/p2p/discv4/lookup.rs
@@ -197,7 +197,7 @@ impl Discv4LookupHandler {
     fn peers_to_ask_push(&self, peers_to_ask: &mut Vec<Node>, target_node_id: H256, node: Node) {
         let distance = bucket_number(target_node_id, node.node_id());
 
-        if peers_to_ask.len() < MAX_NODES_PER_BUCKET {
+        if (peers_to_ask.len() as u64) < MAX_NODES_PER_BUCKET {
             peers_to_ask.push(node);
             return;
         }
@@ -247,7 +247,7 @@ impl Discv4LookupHandler {
             match tokio::time::timeout(Duration::from_secs(5), request_receiver.recv()).await {
                 Ok(Some(mut found_nodes)) => {
                     nodes.append(&mut found_nodes);
-                    if nodes.len() == MAX_NODES_PER_BUCKET {
+                    if nodes.len() as u64 == MAX_NODES_PER_BUCKET {
                         return Ok(nodes);
                     };
                 }
